B-2 / F-16 Fighter Jet Game.
A fighter jet / bomber game I made near the end of March after the invasion of Crimea on March 5th.
It's lotsa fun. The Driver file is the main file that calls all the functions to run the game. Be sure to include the other .cpp files in the same directory.

Pick your plane and go get 'em, tiger.